SolarChem

Documentation for SolarChem.

SolarChem.calcweights

SolarChem.estimateuncertainty

SolarChem.calcweights — Method

calcweights(v ; weightings)

Calculate weightings for a population v for a non-biased sampling of the population. Optionally give weightings as a Dict of "type" => x to further suppress values of "type" by a factor of x.

e.g. julia> v = ["a", "b", "a"] julia> calcweights(v) 3-element Vector{Float64}: 0.5 1.0 0.5

SolarChem.estimateuncertainty — Method

estimateuncertainty(d, unc; maxpctunc, minuncs)

Calculate an assumed uncertainty unc (in %) for all analytes in NamedTuple d. If there are > minuncs (30 by default) uncertainties corresponding to measurements of an analyte, it calculates unreported uncertainties as the mean relative uncertainty. Replaces any values exceeding the maximum percent uncertainty (maxpctunc) with unc.

Requires that uncertainties are given as 1σ with keys :sX where X is the analyte name (e.g. :Na and :sNa).