SolarChem

Documentation for SolarChem.

SolarChem.metadata

SolarChem.periodictable

SolarChem.strelements

SolarChem.bootstrapmean

SolarChem.bootstrapresample

SolarChem.calcweights

SolarChem.estimateuncertainty

SolarChem.pulltopic

SolarChem.trimnans

SolarChem.metadata — Constant

metadata = (:sample,:type,:group)

All metadata names in the NamedTuple generated from Astromat data. Still in development and eventually will become a constant in the SolarChem module.

SolarChem.periodictable — Constant

SolarChem.periodictable

Tuple containing all elements (as Symbols) in order of increasing Z, up to Z=96. A constant within the scope of the SolarChem module.

SolarChem.strelements — Constant

strelements

Tuple containing all elements in order of increasing Z (up to Z=96) as a String followed by a space. Derived from SolarChem.periodictable for use inside loadastromatdata. A constant within the scope of the SolarChem module.

SolarChem.bootstrapmean — Method

bootstrapmean(n, x, σ, weights; rng)

Returns a Vector of n means, each calculated from a random resampling of dataset x with (normally distributed) 1σ uncertainties σ, given weights (obtained from calcweights). If you desire unweighted resamplings provide ones(length(x)) for weights.

Optionally provide a (pseudo)random number generator rng (default: Xoshiro256++).

SolarChem.bootstrapresample — Method

bootstrapresample(n, x, σ, weights; rng)

Returns a Vector of n random samples from dataset x with (normally distributed) 1σ uncertainties σ, given weights (obtained from calcweights). If you desire an unweighted resample provide ones(length(x)) for weights.

Optionally provide a (pseudo)random number generator rng (default: Xoshiro256++).

SolarChem.calcweights — Method

calcweights(v ; weightings)

Calculate weightings for a population v for a non-biased sampling of the population. Optionally give weightings as a Dict of "type" => x to further suppress values of "type" by a factor of x.

e.g. julia> v = ["a", "b", "a"] julia> calcweights(v) 3-element Vector{Float64}: 0.5 1.0 0.5

SolarChem.estimateuncertainty — Method

estimateuncertainty(d, unc; maxpctunc, minuncs)

Calculate an assumed uncertainty unc (in %) for all analytes in NamedTuple d. Replaces any values exceeding the maximum percent uncertainty (maxpctunc) with unc. If there are > minuncs (30 by default) uncertainties corresponding to measurements of an analyte, it calculates unreported uncertainties as the mean relative uncertainty of all measurements (with %unc < maxpctunc).

Requires that uncertainties are given as 1σ with keys :sX where X is the analyte name (e.g. :Na and :sNa).

SolarChem.pulltopic — Method

pulltopic(data, name, topic[s]; exactmatch=true)

Given a NamedTuple of data, returns a NamedTuple with all the names of data and only including those rows within under name (::Symbol) that contain the provided topic(s) –- as a String or Tuple of Strings.

SolarChem.trimnans — Method

trimnans(data, names; alsoinclude = SolarChem.metadata)

Given a NamedTuple of data, return only the values and uncertainties of all non-NaN instances of name –- a name (::Symbol) or Tuple of names from data. Also returns the the Vectors corresponding to the fields listed in alsoinclude, which is by definition those defined by SolarChem.metadata.