SolarChem

Documentation for SolarChem.

SolarChem.oxideconversion

SolarChem.strelements

SolarChem.assigngroup

SolarChem.assigntype

SolarChem.bootstrapelements

SolarChem.bootstrapratios

SolarChem.bsmean

SolarChem.bsmean!

SolarChem.bsresample

SolarChem.bsresample!

SolarChem.calcweights

SolarChem.countmeasurements

SolarChem.countratios

SolarChem.estimateuncertainty

SolarChem.exclude

SolarChem.excludeheated

SolarChem.innergroups

SolarChem.majors

SolarChem.metadata

SolarChem.outergroups

SolarChem.periodictable

SolarChem.pullgroup

SolarChem.pulltopic

SolarChem.pulltype

SolarChem.readvalue

SolarChem.trimextremes

SolarChem.trimnans

SolarChem.unitconversionfactor

SolarChem.oxideconversion — Constant

oxideconversion

A NamedTuple with names of elements corresponding to a multiplicative factor to convert oxide composition to elemental composition. Fe corresponds to ferrous iron (FeO) and Fe3 to ferric iron (Fe₂O₃). oxideconversion is a const within the scope of the SolarChem module.

CaO * oxideconversion.Ca = Ca

Ca / oxideconversion.Ca = CaO

SolarChem.strelements — Constant

SolarChem.strelements

A Tuple containing all elements in order of increasing Z (up to Z=96) as a String followed by a space. Derived from SolarChem.periodictable for use inside loadastromatdata. A const within the scope of the SolarChem module.

SolarChem.assigngroup — Method

assigngroup(str)

Assign all corresponding genetic groups to a String str containing meteorite group/type information. See innergroups and outergroups for groups.

SolarChem.assigntype — Method

assigntype(str)

Assign all corresponding genetic groups to a String `str` containing meteorite group/type information. See [`innergroups`](@ref) and [`outergroups`](@ref) for groups.

SolarChem.bootstrapelements — Method

bootstrapelements(n::Int, data::NamedTuple, elements; resamplemeans=true, weighted=true, rng)

Returns a NamedTuple of vectors of n bootstrap resampled data for each element in the Tuple elements.

Resamples Monte Carlo'ed means by default. Declare resamplemeans=false to return resampled values.

By default weights resampling by sample abundance (based on occurences of unique meteorite names (field :name in data). To remove weighting, declare weighted=false.

see also: trimnans, calcweights, bsresample, bsmean

SolarChem.bootstrapratios — Method

bootstrapratios(n::Int, data::NamedTuple, elements, divisor::Symbol; resamplemeans=true, weighted=true, rng)

Returns a NamedTuple of vectors of n bootstrap resampled data for each element in the Tuple elements, as a ratio of element/divisor (calculated prior to resampling).

Resamples Monte Carlo'ed means by default. Declare resamplemeans=false to return resampled values.

By default weights resampling by sample abundance (based on occurences of unique meteorite names (field :name in data). To remove weighting, declare weighted=false.

see also: trimnans, calcweights, bsresample, bsmean

SolarChem.bsmean! — Method

bsmean!(v, x, σ; w, rng)

In-place version of bsmean that takes a vector v to fill with resampled means.

see also: bsmean

SolarChem.bsmean — Method

bsmean(n, x, σ; w, rng)

Returns a Vector of n means, each calculated from a random resampling (with replacement) of dataset x with (normally distributed) 1σ uncertainties σ. Optionally provide weights (obtained from calcweights, unweighted by default) and a (pseudo)random number generator rng (default: Xoshiro256++).

SolarChem.bsresample! — Method

bsresample!(v, x, σ; w, rng)

In-place version of bsresample that takes a vector v to fill with resamples.

see also: bsmean

SolarChem.bsresample — Method

bsresample(n, x, σ; w, rng)

Returns a Vector of n random samples from dataset x with (normally distributed) 1σ uncertainties σ. Optionally provide weights w (obtained from calcweights, unweighted by default) and a (pseudo)random number generator rng (default: Xoshiro256++).

SolarChem.calcweights — Method

calcweights(v ; weightings)

Calculate weightings for a population v for a non-biased sampling of the population. Optionally give weightings as a Dict of "type" => x to further suppress values of "type" by a factor of x.

e.g. julia> v = ["a", "b", "a"] julia> calcweights(v) 3-element Vector{Float64}: 0.5 1.0 0.5

SolarChem.countmeasurements — Method

countmeasurements(d::NamedTuple, element::Symbol)

Returns the total number of measurements for a given element in d.

countmeasurements(d::NamedTuple, elements::Tuple)

Returns a NamedTuple with names in elements and the total number of measurements for each member of elements in d.

SolarChem.countratios — Method

countratios(d::NamedTuple, numerator::Symbol, denomenator::Symbol)

Returns the total number of non-NaN ratios of numerator/denomenator in d.

countmeasurements(d::NamedTuple, numerators::Tuple, denomenator::Symbol)

Returns a NamedTuple with names in numerators and the total number of non-NaN ratios in d for each ratio of numerators ./ denomenator.

SolarChem.estimateuncertainty — Method

estimateuncertainty(d, unc; uncextrema, minuncs)

Calculate an assumed uncertainty unc (in %) for all analytes in NamedTuple d. Replaces any values outside the bounds of the minimum and maximum percent uncertainty (uncextrema= (0.01%, 50%) by default) with unc. If there are > minuncs (30 by default) uncertainties corresponding to measurements of an analyte, it calculates unreported uncertainties as the mean relative uncertainty of all measurements within uncextrema.

Requires that uncertainties are given as 1σ with keys :sX where X is the analyte name (e.g. :Na -> :sNa).

Using estimateuncertainty! overwrites d.

SolarChem.exclude — Method

exclude(data::NamedTuple, name::Symbol, s)

Exclude all indices containing string (or Tuple of strings) s from the specified name in data.

SolarChem.excludeheated — Method

excludeheated(d::NamedTuple)

Exclude data associated with heating experiments from d. Removes all rows with comments containing °, degrees, 0 C, and 5 C.

see also: exclude

SolarChem.innergroups — Method

innergroups()

Return a Tuple containing all chondrite genetic groups associated with the inner solar system and used in SolarChem.

SolarChem.majors — Method

majors()

Return a Tuple containing major elements.

SolarChem.metadata — Method

metadata = (:name,:type,:group)

Return a Tuple of metadata names in the NamedTuple generated from Astromat data. Still in development.

SolarChem.outergroups — Method

outergroups()

Return a Tuple containing all chondrite genetic groups associated with the outer solar system and used in SolarChem.

SolarChem.periodictable — Method

SolarChem.periodictable()

Return a Tuple containing all elements (as Symbols) in order of increasing Z, up to Z=96.

SolarChem.pullgroup — Method

pullgroup(data, group; exactmatch=false)

Given a NamedTuple of chondrite data, returns a NamedTuple with all the names of data, including only the rows corresponding to the provided group(s) –- as a Symbol or Tuple of Symbols.

SolarChem.pulltopic — Method

pulltopic(data, name, topic[s]; exactmatch=true)

Given a NamedTuple of data, returns a NamedTuple with all the names of data and only including those rows within under name (::Symbol) that contain the provided topic(s) –- as a String or Tuple of Strings.

SolarChem.pulltype — Method

pulltype(data, type; exactmatch=false)

Given a NamedTuple of chondrite data, returns a NamedTuple with all the names of data, including only the rows corresponding to the provided type(s) –- as an Integer or Tuple of Integers.

SolarChem.readvalue — Method

readvalue(x)

Return a value xthat is either aNumberas a float or parsesxif it's a String. An empty String yieldsNaN`. Tolerates usage of commas in numbers.

SolarChem.trimextremes — Method

trimextremes(data; min=0, max=1)

Cycle through all elements in data and replace any value in excess of min and max (set by default to the physical limits of fractional mass abundance) to a NaN.

Using trimextremes! overwites data.

SolarChem.trimnans — Method

trimnans(data, names; alsoinclude = SolarChem.metadata)

Given a NamedTuple of data, return only the values and uncertainties of all non-NaN instances of name –- a name (::Symbol) or Tuple of names from data. Also returns the the Vectors corresponding to the fields listed in alsoinclude, which is by definition those defined by SolarChem.metadata.

SolarChem.unitconversionfactor — Method

unitconversionfactor(str)

Returns a unit conversion factor to scale measurements by to convert to mass fraction abundance based on the unit given in the heading str.

includes: wt% (%, %m/m), mg/g, μg/g (ppm), ng/g (ppb), pg/g excludes: at%, vol%