SolarChem

Documentation for SolarChem.

SolarChem.oxideconversion

SolarChem.strelements

SolarChem.Composition

SolarChem.Fractions

SolarChem.LoddersFegley1998

SolarChem.allsolar

SolarChem.assigngroup

SolarChem.assigntype

SolarChem.astromatdata

SolarChem.bootstrapelements

SolarChem.bootstrapratios

SolarChem.bsmean

SolarChem.bsmean!

SolarChem.bsresample

SolarChem.bsresample!

SolarChem.calcweights

SolarChem.countmeasurements

SolarChem.countnotnans

SolarChem.countratios

SolarChem.estimateuncertainty

SolarChem.exclude

SolarChem.excludeheated

SolarChem.fastll

SolarChem.fraction2ratio

SolarChem.fraction2ratio!

SolarChem.heavysolar

SolarChem.innergroups

SolarChem.jump

SolarChem.lightsolar

SolarChem.lingce

SolarChem.lldist

SolarChem.loadastromatdata

SolarChem.majors

SolarChem.mediumsolar

SolarChem.metadata

SolarChem.normpdf

SolarChem.outergroups

SolarChem.periodictable

SolarChem.pullgroup

SolarChem.pulltopic

SolarChem.pulltype

SolarChem.quadgce

SolarChem.ratio2fraction

SolarChem.ratiosummary

SolarChem.readvalue

SolarChem.removefrom

SolarChem.solarlogmix

SolarChem.solarmixmetropolis

SolarChem.solartwins

SolarChem.sun

SolarChem.trimextremes

SolarChem.trimnans

SolarChem.unitconversionfactor

SolarChem.unweightedmean

SolarChem.weightedmean

SolarChem.oxideconversion — Constant

oxideconversion

A NamedTuple with names of elements corresponding to a multiplicative factor to convert oxide composition to elemental composition. Fe corresponds to ferrous iron (FeO) and Fe3 to ferric iron (Fe₂O₃). oxideconversion is a const within the scope of the SolarChem module.

CaO * oxideconversion.Ca = Ca

Ca / oxideconversion.Ca = CaO

SolarChem.strelements — Constant

SolarChem.strelements

A Tuple containing all elements in order of increasing Z (up to Z=96) as a String followed by a space. Derived from SolarChem.periodictable for use inside SolarChem.loadastromatdata. A const within the scope of the SolarChem module.

SolarChem.Composition — Type

Composition

Custom struct that contains a chemical composition, represented as the normally distributed mean m and standard deviation s of the measured/modeled composition.

SolarChem.Fractions — Type

Fractions(outer, sun)

Custom struct with proposed mixture parameters: the fractional contribution of outer solar system material (inner solar system contribution = 1 - outer) to the solar budget and the mass fraction of the sun worth of chondritic material added to the solar photosphere.

SolarChem.LoddersFegley1998 — Method

LoddersFegley1998()

Returns a NamedTuple containing the meteorite chemistry data reported in Tables 2.3, 16.10, and 16.11 of Lodders & Fegley (1998, The Planetary Scientist's Companion, Oxford University Press).

Fields include element names (els), atomic number (z), 50% condensation temperature (Tc), inner solar system mean composition (comprised of EH, EL, H, L, LL, and R groups), and outer solar system mean composition (comprised of CI, CM, CV, CO, CK, and CR groups).

SolarChem.allsolar — Method

allsolar(; Fe=true)

Return a Tuple of all elements with solar twin data reported in Bedell+ 2018 (ApJ, doi:10.3847/1538-4357/aad908). Excludes :Fe by default, but Fe=true will include it.

see also: lightsolar, mediumsolar, heavysolar

SolarChem.assigngroup — Method

assigngroup(str)

Assign all corresponding genetic groups to a String str containing meteorite group/type information. See innergroups and outergroups for groups.

SolarChem.assigntype — Method

assigntype(str)

Assign all corresponding genetic groups to a String `str` containing meteorite group/type information. See [`innergroups`](@ref) and [`outergroups`](@ref) for groups.

SolarChem.astromatdata — Method

    astromatdata(; url, datasets_per_request=1000, save=true, showskipped=false)

Get elemental geochemical data (only) from the AstroMat database for all meteorites, returned as a NamedTuple. The default url string is:

"https://api.astromat.org/v4/search/results?analysisTypes=rock%3A%3A[WHOLE+ROCK]&taxons=METEORITE&variables=MAJ||REE||TE||VO"

This requests whole rock major (MJ), volatile (VO), trace (TE), and rare earth element (REE) data for all meteorites in the database. You may provide your own customized API string, but this is not recommended unless you are familiar with the AstroMat API.

By default, the function requests 1000 datasets_per_request, which is a roughly optimized amount.

When save is set to true or provided with a filepath, the data is serlialized to the provided filepath or to a file automatically labeled astromat-YYYY-MM-DD.jls (reflecting the current date). If false is provided, the output is not saved.

When showskipped is true, returns a tuple of the data from astromat, as well as a list of excluded variable names.

SolarChem.bootstrapelements — Method

bootstrapelements(n::Int, data::NamedTuple, elements; resamplemeans=true, weighted=true, rng)

Returns a NamedTuple of vectors of n bootstrap resampled data for each element in the Tuple elements.

Resamples Monte Carlo'ed means by default. Declare resamplemeans=false to return resampled values.

By default weights resampling by sample abundance (based on occurences of unique meteorite names (field :name in data). To remove weighting, declare weighted=false.

see also: trimnans, calcweights, bsresample, bsmean

SolarChem.bootstrapratios — Method

bootstrapratios(n::Int, data::NamedTuple, elements, divisor::Symbol; resamplemeans=true, fractional=false, weighted=true, rng)

Returns a NamedTuple of vectors of n bootstrap resampled data for each element in the Tuple elements, as a ratio of element/divisor (calculated prior to resampling).

Resamples Monte Carlo'ed means by default. Declare resamplemeans=false to return resampled values.

fractional=true resamples ratios as fractions $\frac{x}{x+y}$ and then algebraically converted to ratios $\frac{x}{y}$. This mitigates resampled outliers stemming from small values of $y$. If fractional=false, ratios are calculated directly.

By default weights resampling by sample abundance (based on occurences of unique meteorite names (field :name in data). To remove weighting, declare weighted=false.

see also: trimnans, calcweights, bsresample, bsmean

SolarChem.bsmean! — Method

bsmean!(v, x, σ; w, rng)

In-place version of bsmean that takes a vector v to fill with resampled means.

see also: bsmean

SolarChem.bsmean — Method

bsmean(n, x, σ; w, rng)

Returns a Vector of n means, each calculated from a random resampling (with replacement) of dataset x with (normally distributed) 1σ uncertainties σ. Optionally provide weights (obtained from calcweights, unweighted by default) and a (pseudo)random number generator rng (default: Xoshiro256++).

SolarChem.bsresample! — Method

bsresample!(v, x, σ; w, rng)

In-place version of bsresample that takes a vector v to fill with resamples.

see also: bsmean

SolarChem.bsresample — Method

bsresample(n, x, σ; w, rng)

Returns a Vector of n random samples from dataset x with (normally distributed) 1σ uncertainties σ. Optionally provide weights w (obtained from calcweights, unweighted by default) and a (pseudo)random number generator rng (default: Xoshiro256++).

SolarChem.calcweights — Method

calcweights(v ; weightings)

Calculate weightings for a population v for a non-biased sampling of the population. Optionally give weightings as a Dict of "type" => x to further suppress values of "type" by a factor of x.

e.g. julia> v = ["a", "b", "a"] julia> calcweights(v) 3-element Vector{Float64}: 0.5 1.0 0.5

SolarChem.countmeasurements — Method

countmeasurements(d::NamedTuple, element::Symbol)

Returns the total number of measurements for a given element in d.

countmeasurements(d::NamedTuple, elements::Tuple)

Returns a NamedTuple with names in elements and the total number of measurements for each member of elements in d.

SolarChem.countnotnans — Method

SolarChem.countnotnans(x)

Count the number of non-NaN elements in x.

SolarChem.countratios — Method

countratios(d::NamedTuple, numerator::Symbol, denomenator::Symbol)

Returns the total number of non-NaN ratios of numerator/denomenator in d.

countmeasurements(d::NamedTuple, numerators::Tuple, denomenator::Symbol)

Returns a NamedTuple with names in numerators and the total number of non-NaN ratios in d for each ratio of numerators ./ denomenator.

SolarChem.estimateuncertainty — Method

estimateuncertainty(d, unc; uncextrema, minuncs)

Calculate an assumed uncertainty unc (in %) for all analytes in NamedTuple d. Replaces any values outside the bounds of the minimum and maximum percent uncertainty (uncextrema= (0.01%, 50%) by default) with unc. If there are > minuncs (30 by default) uncertainties corresponding to measurements of an analyte, it calculates unreported uncertainties as the mean relative uncertainty of all measurements within uncextrema.

Requires that uncertainties are given as 1σ with keys :sX where X is the analyte name (e.g. :Na -> :sNa).

Using estimateuncertainty! overwrites d.

SolarChem.exclude — Method

exclude(data::NamedTuple, name::Symbol, s)

Exclude all indices containing string (or Tuple of strings) s from the specified name in data.

SolarChem.excludeheated — Method

excludeheated(d::NamedTuple)

Exclude data associated with heating experiments from d. Removes all rows with comments containing °, degrees, 0 C, and 5 C.

see also: exclude

SolarChem.fastll — Method

fastll(x, m, s)

Calculate the relative log-likelihood that x is drawn from the Normal distribution m ± s. fastll excludes the additive constant -log(sqrt(2π)*σ), which can be multiplied to summations of exp(fastll(..)) later for accurate likelihood calculations. All inputs must be Float64s.

SolarChem.fraction2ratio! — Method

fraction2ratio!(x::Vector)

Convert a Vector of fractions x to ratios, as by fraction2ratio.

SolarChem.fraction2ratio — Method

fraction2ratio(x)

Algebraicaly convert a fraction x of the form $\frac{a}{a+b}$ to a ratio of the form $\frac{a}{b}$.

SolarChem.heavysolar — Method

heavysolar()

Return a Tuple of Period 5-6 elements with solar twin data reported in Bedell+ 2018 (ApJ, doi:10.3847/1538-4357/aad908).

see also: allsolar

SolarChem.innergroups — Method

innergroups()

Return a Tuple containing all chondrite genetic groups associated with the inner solar system and used in SolarChem.

SolarChem.jump — Method

jump(p::Fractions, j::Fractions; rng)

Perturb a random field of p (outer,sun) with a random gaussian jump σ described by the corresponding field in j. Ensures that the new proposal satisfies that 0 < outer < 1 and 0 < sun < 1. Optionally supply a random number generation seed rng.

see also: Fractions

SolarChem.lightsolar — Method

lightsolar()

Return a Tuple of Period 1-3 elements with solar twin data reported in Bedell+ 2018 (ApJ, doi:10.3847/1538-4357/aad908).

see also: allsolar

SolarChem.lingce — Method

SolarChem.lingce(m, b, age, x)

Recalculate the composition of element x for a star of a given age, correcting for galactic chemical evolution (GCE) with a linear regression of slope m and intercept b.

SolarChem.lldist — Method

lldist(μ, σ, μₒ, σₒ; n=100)

Calculates the log-likelihood that the normally distributed measurement(s) μ ± σ was/were drawn from the model composition μₒ ± σₒ, integrated over n nodes of the 3σ limits of the model distribution.

Accepts discrete Float64 values or vectors of Float64s for μ and σ. Values of NaN in modeled data are ignored and do not contribute to log-likelihood calculations.

see also: normpdf

SolarChem.loadastromatdata — Function

    SolarChem.loadastromatdata(file::String)

DEPRECATED: use astromatdata, which directly accesses the AstroMat API.

Load data exported from Astromat as a csv. Given underlying group and type assignment functions, this will only load chondrite data.

see also: assigngroup, assigntype, unitconversionfactor, readvalue

SolarChem.majors — Method

majors()

Return a Tuple containing major elements.

SolarChem.mediumsolar — Method

mediumsolar(; Fe=true)

Return a Tuple of Period 4 elements with solar twin data reported in Bedell+ 2018 (ApJ, doi:10.3847/1538-4357/aad908).

Excludes :Fe by default, but Fe=true will include it.

see also: allsolar

SolarChem.metadata — Method

metadata = (:name,:type,:group)

Return a Tuple of metadata names in the NamedTuple generated from Astromat data. Still in development.

SolarChem.normpdf — Method

normpdf(x, m, s)

Calculate the probability density of x given a normal distribution with mean m and standard deviation s. All inputs must be Float64s.

$\mathcal{L}(\mu \pm \sigma | x ) = f(x | \mu \pm \sigma) = \frac{e^{-\frac{1}{2} \left( \frac{x-\mu}{\sigma}\right)}}{\sigma\sqrt{2\pi}}$

SolarChem.outergroups — Method

outergroups()

Return a Tuple containing all chondrite genetic groups associated with the outer solar system and used in SolarChem.

SolarChem.periodictable — Method

SolarChem.periodictable()

Return a Tuple containing all elements (as Symbols) in order of increasing Z, up to Z=96.

SolarChem.pullgroup — Method

pullgroup(data, group; exactmatch=false)

Given a NamedTuple of chondrite data, returns a NamedTuple with all the names of data, including only the rows corresponding to the provided group(s) –- as a Symbol or Tuple of Symbols.

SolarChem.pulltopic — Method

pulltopic(data, name, topic[s]; exactmatch=true)

Given a NamedTuple of data, returns a NamedTuple with all the names of data and only including those rows within under name (::Symbol) that contain the provided topic(s) –- as a String or Tuple of Strings.

SolarChem.pulltype — Method

pulltype(data, type; exactmatch=false)

Given a NamedTuple of chondrite data, returns a NamedTuple with all the names of data, including only the rows corresponding to the provided type(s) –- as an Integer or Tuple of Integers.

SolarChem.quadgce — Method

SolarChem.quadgce(a, b, age, x)

Recalculate the composition of element x for a star of a given age, correcting for galactic chemical evolution (GCE) using a quadratic regression with parameters a and b of the form: $at + bt^2$

SolarChem.ratio2fraction — Method

fraction2ratio(x)

Algebraicaly convert a ratio x of the form $\frac{a}{b}$ to a fraction of the form $\frac{a}{a+b}$.

SolarChem.ratiosummary — Method

ratiosummary(resampled_measurements, resampled_ratios; ratiocounts, minratios)

Calculate ratio mean and standard deviation (1σ) from resampled_measurements calculated by bootstrapelements and resampled_ratios calculated bybootstrapratios.

If no ratios were calculated for a field in resampled_ratios (all values NaN), μ/σ are calculated from the corresponding resampled_measurements.

Optionally, provide ratiocounts calculated by countratios and a minimum number of ratios minratios for each element required to use resampled_ratios. If the ratio counts for an element is less than minratios, μ/σ are calculated from the corressponding elements in resampled_measurements.

NamedTuple(x) for some x = ratiosummary... will convert the result to a NamedTuple of Compositions.

SolarChem.readvalue — Method

readvalue(x)

Return a value x that is either a Number as a float or parses x if it's a String. An empty String yields NaN. Tolerates usage of commas in numbers.

SolarChem.removefrom — Method

removefrom(s, x)

Remove key s and corresponding value from NamedTuple x. Alerts user if s is not present in x.

julia> removefrom(:c, (; a=1, b=2, c=3))
(a = 1, b = 2)

julia> removefrom(:c, (; a=1, b=2))
`:c` is not present in this NamedTuple
(a = 1, b = 2)

SolarChem.solarlogmix — Method

solarlogmix(inner, outer, solar, f::Fractions; solarunc=true)

Calculate the solar photosphere-normalized composition of a model Sun that ingests f.sun M⊙ of chondritic material comprised of f.outer proportion of outer solar system composition and 1 - f.outer proportion of inner solar system composition. The args inner, outer, and solar are all instances of a Composition struct.

The result is reported in dex, as the ratio of log-ratios: log10(X/D) - log10(X⊙/D⊙) where D denotes the ratio divisor (e.g. Fe, H) and ⊙ denotes solar compositions.

SolarChem.solarmixmetropolis — Method

solarmixmetropolis...

Metropolis algorithm to estimate the requisite mixture of chondritic components added to the solar photosphere to match it to solar twin compositions.

SolarChem.solartwins — Method

solartwins(; gce=true, includesun=false)

Returns a named tuple containing stellar parameters and solar-normalized compositions ([X/Fe], in dex) either corected for galactic chemical evolution (gce=true) or left uncorrected (gce=false). All corresponding uncertainty fields are preceded with s, e.g. :Ca and :sCa or logg and slogg. By default excludes the Sun (improves mixing model efficiency), but you can include solar data (mostly 0 dex by definition) with includesun=true.

Data is compiled from Bedell + 2018 (ApJ, doi:10.3847/1538-4357/aad908). Where multiple species are reported (e.g. CI-CH, ScI-ScII, TiI-TiII, CrI-CrII), this reports the mean and quadrature-propagated uncertainties of the two species.

SolarChem.sun — Function

sun(divisor=nothing; logspace=true)

Returns a NamedTuple containing the solar photosphere composition, as reported by Asplund+ 2021 [A&A, doi:10.1051/0004-6361/202140445]. Optionally provide an element name for divisoras a Symbol, to return the divisor-normalized ratios with uncertainties propagated in quadrature. logspace=false converts values and uncetainties to linear-space ratios.

SolarChem.trimextremes — Method

trimextremes(data; min=0, max=1)

Cycle through all elements in data and replace any value in excess of min and max (set by default to the physical limits of fractional mass abundance) to a NaN.

Using trimextremes! overwites data.

SolarChem.trimnans — Method

trimnans(data, names; alsoinclude = SolarChem.metadata)

Given a NamedTuple of data, return only the values and uncertainties of all non-NaN instances of name –- a name (::Symbol) or Tuple of names from data. Also returns the the Vectors corresponding to the fields listed in alsoinclude, which is by definition those defined by SolarChem.metadata.

SolarChem.unitconversionfactor — Method

unitconversionfactor(str; io=stdout)

Returns a unit conversion factor to scale measurements by to convert to mass fraction abundance based on the unit given in the heading str. By default, notifies user if a unit is ignored.

includes: wt% (%, %m/m), mg/g, μg/g (ppm), ng/g (ppb), pg/g excludes: at%, vol%, per mil, Ma, etc...

SolarChem.unweightedmean — Method

unweightedmean(m, s; sem=false)

Given values m (e.g. measurement means) with absolute 1σ uncertainty s, returns the unweighted mean and 1σ uncertainty estimated by addition in quadrature.

$\sigma = \sqrt{\sum^n_{i=1} s_i^2}$

If sem=true, the 1σ uncertainty is the standard error of the mean ($\sigma_{\bar x} = \sigma/\sqrt{n}$).

SolarChem.weightedmean — Method

weightedmean(m, s)

Given values m (e.g. measurement means) with absolute 1σ uncertainties s, returns the variance-weighted mean and 1σ uncertainty.

$\mu = \frac{\sum^n_{i=1} x_i~\sigma_i^{-2}}{\sum^n_{i=1}\sigma_i^{-2}}$

$\sigma = \sqrt{\frac{1}{\sum^n_{i=1}\sigma_i^{-2}}}$